Create pf_api_min_1_1.js
Added:
Appending this on to the link sets the "Business Unit" search criteria (see "Business Unit" tab for countries and id's)
&com.peopleclick.cp.formdata.FLD_JP_DEPARTMENT=1454

On the test example the PF server returns error (wrong query). The parameter is correctly attached as in documentation. Please test only on updated PF API servers to see the correct results.